Begin big rework of app logic
I was doing all my work in the Angular controller that was created to manage
the thumbnail list in the top ribbon. But all of the UI widgets are living
in the side pane, which was a child of the LayerManagerController. This
was a bit silly.

So, the contoller I had is now called the StarHuntThumbsController, and the
StarHuntController is now attached to the left pane full of widgets. This
should be a lot easier to use going forward.