Add a "report" of the coordinates of some of the markers
There are now two kinds of markers, "coordinate markers" and regular ones.
There is a new button, "Report Coordinates", that will open a modal dialog box
that prints out the coordinates of the "coordinate markers" and provides a
button to copy them to the clipboard. The coordinates are printed out to
second/arcsecond precision.